fix(pwa-tools): normalize path segment separator in paths
Previosuly, the OS-specific separator was used (i.e. `/` on *nix, `\` on
Windows). This made it more difficult (and unintuitive) to match URLs (e.g. for
`ngsw-manifest.json#static.ignore`), because one had to account for the fact
that the URLs might look different, based on the OS the manifest was build on.

Before:
```json
"static.ignore": [
  "^(?:/|\\\\)please(?:/|\\\\)ignore(?:/|\\\\)this(?:/|\\\\)dir(?:/|\\\\)",
]

After:
```json
"static.ignore": [
  "^/please/ignore/this/dir/",
]
```